tests/Gateway.Api.Tests/Unit/ProcessControllerTest.cs
using FluentAssertions;
using Gateway.Api.Controllers;
using Gateway.Common.Commands;
using Microsoft.AspNetCore.Http;
using Microsoft.AspNetCore.Mvc;
using Moq;
using NUnit.Framework;
using RawRabbit;
using System;
using System.Collections.Generic;
using System.Security.Claims;
using System.Text;
using System.Threading.Tasks;

namespace Gateway.Api.Tests
{
    public class ProcessControllerTests
    {
        [Test]
        public void home_controller_get_should_return_string_content()
        {
            var busClientMock = new Mock<IBusClient>();
            var controller = new ProcessController(busClientMock.Object);

            var result = controller.Get();

            var contentResult = result as ContentResult;
            contentResult.Should().NotBeNull();
            contentResult.Content.Should().BeEquivalentTo("Hello from Process Controller API!");
        }

        [Test]
        public async Task activities_controller_post_should_return_accepted()
        {
            var busClientMock = new Mock<IBusClient>();
            var controller = new ProcessController(busClientMock.Object);
            var userId = Guid.NewGuid();
            controller.ControllerContext = new ControllerContext
            {
                HttpContext = new DefaultHttpContext
                {
                    User = new ClaimsPrincipal(new ClaimsIdentity(new Claim[]
                        {
                            new Claim(ClaimTypes.Name, userId.ToString())
                        }, "test"))
                }
            };

            var command = new ProcessPayment
            {
                UserId = userId
            };

            var result = await controller.Post(command);

            var contentResult = result as AcceptedResult;
            contentResult.Should().NotBeNull();
            contentResult.Location.Should().StartWith($"processes/");
        }
    }
}
